Requirement fix
Fixed the code so that it aligns with the requirement of the assignment. The QA tree is now only loaded once at the beginning of the program and only stored once when the player stops playing. Also added a message telling the user of wrongful input when answering not answering a question with yes or no